Fix 'scala.throws' symbol lookup in the compiler plugin.
The use of typeOf[] triggered MissingRequirementErrors when running as a
plugin and not loaded in the target classpath; object coop.plausible.scala.nx.NX
could not be found in the compiler mirror.

Replacing typeOf[] with rootMirror.staticClass allows us to explicitly
avoid target runtime dependnencies on the NX classes from the compiler
plugin.